// igamt-hl7-client-v2/src/app/modules/shared/constants/type.enum.ts
export enum Type {
  IGDOCUMENT = 'IGDOCUMENT',
  DATATYPEREGISTRY = 'DATATYPEREGISTRY',
  DATATYPELIBRARY = 'DATATYPELIBRARY',
  DERIVEDDATATYPEREGISTRY = 'DERIVEDDATATYPEREGISTRY',
  SEGMENTREGISTRY = 'SEGMENTREGISTRY',
  VALUESETREGISTRY = 'VALUESETREGISTRY',
  PROFILECOMPONENTREGISTRY = 'PROFILECOMPONENTREGISTRY',
  COMPOSITEPROFILEREGISTRY = 'COMPOSITEPROFILEREGISTRY',
  COCONSTRAINTGROUPREGISTRY = 'COCONSTRAINTGROUPREGISTRY',
  DATATYPE = 'DATATYPE',
  VALUESET = 'VALUESET',
  CONFORMANCEPROFILE = 'CONFORMANCEPROFILE',
  SEGMENT = 'SEGMENT',
  PROFILECOMPONENT = 'PROFILECOMPONENT',
  COMPOSITEPROFILE = 'COMPOSITEPROFILE',
  COCONSTRAINTGROUP = 'COCONSTRAINTGROUP',
  SEGMENTREF = 'SEGMENTREF',
  GROUP = 'GROUP',
  FIELD = 'FIELD',
  COMPONENT = 'COMPONENT',
  SUBCOMPONENT = 'SUBCOMPONENT',
  TEXT = 'TEXT',
  PROFILE = 'PROFILE',
  CONFORMANCEPROFILEREGISTRY = 'CONFORMANCEPROFILEREGISTRY',
  DISPLAY = 'DISPLAY',
  EVENT = 'EVENT',
  EVENTS = 'EVENTS',
  BINDING = 'BINDING',
  SECTION = 'SECTION',
  DYNAMICMAPPING = 'DYNAMICMAPPING',
  CONFORMANCESTATEMENTSUMMARY = 'CONFORMANCESTATEMENTSUMMARY',
  CUSTOM_STRUCTURE_EDITOR = 'CUSTOM_STRUCTURE_EDITOR',
  MESSAGESTRUCT = 'MESSAGESTRUCT',
  SEGMENTCONTEXT = 'SEGMENTCONTEXT',
  MESSAGECONTEXT = 'MESSAGECONTEXT',

}